Changes made to port to new version of python (python 3.12.7)
- Needed to change 'print' to 'print()'
- Needed to change 'except Exception, e:' to 'except Exception(e):' etc.

Testing:
It runs on Ubuntu 24.10, with venv, and modbus-tk library with one other
change (see below).

In base directory:
  pyhton -m venv venv
  . venv/bin/activate
  pip install modbus-py
  # Fix 'modbus-tk' (see below)

  cd modsim
  python modsim.py -p 1502 mbmap_test_device.xml

To be about to be run as a regular user, the port needs to be changed to be
above port 1000 (eg 502 -> 1502).

Change to 'modbus-tk'
Change line: venv/lib/python3.12/site-packages/modbus_tk/modbus.py:436
  self._data = [0] * size
to
  self._data = [0] * int(size)